requirements.txt
# Install with: `pip install -r requirements.txt`

Django >=1.8.0, < 1.9
pymongo                         #
mongoengine                     #
djangorestframework == 3.1.0
django-rest-framework-mongoengine == 2.0.1
requests
markdown